+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - This ensures correct resource loading in environments like Spring Boot where the thread context classloader is used for resource loading.
 - Improve low memory breadcrumb capturing ([#4325](https://github.com/getsentry/sentry-java/pull/4325))
 - Fix do not initialize SDK for Jetpack Compose Preview builds ([#4324](https://github.com/getsentry/sentry-java/pull/4324))
+- Fix Synchronize Baggage values ([#4327](https://github.com/getsentry/sentry-java/pull/4327))
 
 ### Improvements